refactor(settings): de-monolith SettingsView (2052 → 85 lines)
SettingsView.swift was the app's clear monolith: all 6 tabs, every
provider's status/plan/connection logic as a repeated 4-way
`switch service` across ~7 helpers, both account sheets, both account
rows, and ~10 reusable primitives — all in one 2052-line file.

Following the UsageDashboardView "R8 split" precedent, this is a pure
move + centralization with no visible behavior change:

- Shell: SettingsView.swift → 85 lines (TabView + fixed window frame +
  the settingsTab scroll/padding wrapper only).
- Per-tab views under Views/Settings/: General, Provider, ApiUsage,
  Cost, About — each owning the shared singletons it observes. The two
  account sheets and the reconnect-failure alert move onto
  ProviderSettingsView, where their controls live.
- Extracted AddClaudeAccountSheet, AddCodexAccountSheet,
  AccountProfileRow (Claude), CodexAccountProfileRow into own files.
- Moved ~10 reusable primitives into Components/SettingsComponents.swift
  (private→internal), plus a new reusable ExtraUsageRow that de-dups the
  pill+manage+detail copy shared by the API-Usage tab and each provider
  pane.
- Centralized the repeated per-provider facts in a new
  ProviderSettingsFacts value type: source/status/plan/access/error/
  statusColor are derived once. ProviderSettingsView gathers each
  provider's live primitives in a single switch and hands them to the
  struct, replacing 7 scattered `switch service` helpers. Folded the
  dead providerStatusColor (defined, never called) in as a tested
  statusColor; moved QuotaBand.severity and String.nilIfEmpty alongside.

Tab order, provider segmented picker, multi-account rendering, both
sheets, and all toggles are unchanged.

Tests: adds ProviderSettingsFactsTests (source/plan/status/statusColor
across all 4 providers + QuotaBand.severity ordering). SettingsView
smoke test is unchanged — the shell still renders the 760×660 layout.
