+++ .agents/SESSIONS/2026-07-14.md
@@ -47,3 +47,58 @@ Claude Code ▸ "Claude Code OAuth usage" (dev builds re-prompt per rebuild).
 `WakeQuotaAuthority`/`LiveWakeQuotaProvider` still call the raw CLI service; the
 session-wake default-account quota gate would also benefit from OAuth, but needs a
 **side-effect-free** OAuth fetch to avoid coupling UI `@Published` state into wake decisions.
+
+---
+
+# 2026-07-14 (pt 2) — De-monolith SettingsView (largest structural chip)
+
+## Problem
+`MeterBar/Views/SettingsView.swift` was 2052 lines: all 6 tabs, every provider's
+status/plan/connection logic as a repeated 4-way `switch service` across ~7
+helpers, both account sheets, both account rows, and ~10 reusable primitives —
+all in one file.
+
+## Refactor (pure move + centralization; no behavior change)
+Followed the UsageDashboardView "R8 split" precedent.
+
+- **Shell:** `SettingsView.swift` → **85 lines** (TabView + fixed window frame +
+  `settingsTab` scroll/padding wrapper only). Owns just `sessionWakeStore` for
+  the feature-flagged Automation tab.
+- **Per-tab views** under `MeterBar/Views/Settings/`: `GeneralSettingsView`,
+  `ProviderSettingsView`, `ApiUsageSettingsView`, `CostSettingsView`,
+  `AboutSettingsView`. Each owns the shared singletons it observes. The two
+  account sheets + the reconnect-failure alert moved onto `ProviderSettingsView`
+  (they're driven entirely by controls in that tab).
+- **Extracted components:** `AddClaudeAccountSheet`, `AddCodexAccountSheet`,
+  `AccountProfileRow` (Claude), `CodexAccountProfileRow` into own files.
+- **Primitives → `Components/SettingsComponents.swift`** (`SettingsInfoRow`,
+  `SettingsPanelSection`, `SettingsRowView`, `SettingsNotice`, `SettingsDivider`,
+  `StatusPill`, `AdminKeySettingsRow`, `SettingsReadonlyField`,
+  `SettingsInput*` modifiers + `View` ext, metrics). `private`→`internal`.
+  Also added a reusable `ExtraUsageRow` that de-dups the pill+manage+detail copy
+  shared by the API-Usage tab and each provider pane.
+- **Centralized provider facts:** new `ProviderSettingsFacts` value type derives
+  source/status/plan/access/error/statusColor **once**. `ProviderSettingsView`
+  gathers each provider's live primitives in a **single** `switch`, hands them to
+  the struct — replacing 7 scattered `switch service` helpers. Killed the dead
+  `providerStatusColor` (was defined, never called) by folding it in as a tested
+  `statusColor`. Moved `QuotaBand.severity` + `String.nilIfEmpty` alongside it.
+
+## Files
+1 shell + 5 tab views + 4 extracted components + 1 view-model + 1 components file
++ 1 test file. `SettingsView.swift` 2052 → 85.
+
+## Tests (TDD for the new accessors)
+`ProviderSettingsFactsTests` — source/plan/status/statusColor across all 4
+providers + `QuotaBand.severity` ordering / worst-band selection.
+`SettingsViewSmokeTests` unchanged (shell still renders the 760×660 tabbed layout).
+
+## Verification
+Local `swift test`/build skipped per MacBook rule — rely on PR CI. Static sweep:
+no duplicate top-level decls, no dangling refs to removed private helpers, all
+referenced symbols resolve. SPM (`path: "MeterBar"`) + Xcode synchronized root
+groups auto-discover the new files — no manifest edits needed.
+
+## Scope
+SettingsView only. Tab order, provider segmented picker, multi-account rendering,
+both sheets, and all toggles unchanged.
